+++ src/index.ts
@@ -6,40 +6,10 @@ import { INotebookTracker } from '@jupyterlab/notebook';
 import { IStatusBar } from '@jupyterlab/statusbar';
 import { ToolbarButton } from '@jupyterlab/apputils';
 import { lockIcon, editIcon } from '@jupyterlab/ui-components';
-import { Widget } from '@lumino/widgets';
 
-class CellLockStatus extends Widget {
-  private _statusNode: HTMLElement;
-  private _timer: number | null = null;
-
-  constructor() {
-    super();
-
-    this._statusNode = document.createElement('span');
-    this._statusNode.classList.add('jp-CellLockStatus');
-    this._statusNode.style.display = 'inline-flex';
-    this._statusNode.style.alignItems = 'center';
-    this._statusNode.style.padding = '0 8px';
-    this._statusNode.style.fontSize = 'var(--jp-ui-font-size1)';
-    this._statusNode.innerText = '';
-    this.node.appendChild(this._statusNode);
-    this.node.style.display = 'inline-flex';
-    this.node.style.alignItems = 'center';
-  }
-
-  setTemporaryStatus(summary: string, timeoutMs = 4000) {
-    this._statusNode.innerText = summary;
-
-    if (this._timer) {
-      window.clearTimeout(this._timer);
-    }
-
-    this._timer = window.setTimeout(() => {
-      this._statusNode.innerText = '';
-      this._timer = null;
-    }, timeoutMs);
-  }
-}
+import { CellLockStatus } from './status';
+import { applyCellLockIcon, refreshLockIcons } from './lockIcon';
+import { toggleCellMetadata } from './metadata';
 
 const plugin: JupyterFrontEndPlugin<void> = {
   id: 'jupyterlab-cell-lock:plugin',
@@ -53,9 +23,6 @@ const plugin: JupyterFrontEndPlugin<void> = {
   ) => {
     console.log('jupyterlab-cell-lock extension activated!');
 
-    const asBool = (v: unknown) => (typeof v === 'boolean' ? v : true);
-
-    // Create a status bar widget
     let statusWidget: CellLockStatus | null = null;
     if (statusBar) {
       statusWidget = new CellLockStatus();
@@ -65,130 +32,12 @@ const plugin: JupyterFrontEndPlugin<void> = {
       });
     }
 
-    const applyCellLockIcon = (
-      cellModel: any,
-      cellWidget: any,
-      retryCount = 0
-    ) => {
-      // JupyterLab may omit "editable"/"deletable" when they are true,
-      // as this is the default. To handle this correctly, the extension treats
-      // missing values as true so the comparison logic works as expected.
-      const editable = asBool(cellModel.getMetadata('editable'));
-      const deletable = asBool(cellModel.getMetadata('deletable'));
-
-      const promptNode = cellWidget.node.querySelector(
-        '.jp-InputPrompt.jp-InputArea-prompt'
-      );
-
-      if (!promptNode) {
-        if (retryCount < 10) {
-          setTimeout(() => {
-            applyCellLockIcon(cellModel, cellWidget, retryCount + 1);
-          }, 10);
-        }
-        return;
-      }
-
-      // Remove old lock icon if present
-      const existing = promptNode.querySelector('.jp-CellLockIcon');
-      if (existing) {
-        existing.remove();
-      }
-
-      // Add the icon if the cell is locked in any way
-      if (!editable || !deletable) {
-        const iconNode = document.createElement('span');
-        iconNode.className = 'jp-CellLockIcon';
-
-        // Construct the tooltip message
-        let tooltipMessage = 'This cell is ';
-        const isReadOnly = !editable;
-        const isUndeletable = !deletable;
-
-        if (isReadOnly && isUndeletable) {
-          tooltipMessage += 'read-only and undeletable.';
-        } else if (isReadOnly) {
-          tooltipMessage += 'read-only but can be deleted.';
-        } else if (isUndeletable) {
-          tooltipMessage += 'undeletable but can be edited.';
-        }
-
-        iconNode.title = tooltipMessage;
-
-        lockIcon.element({
-          container: iconNode,
-          elementPosition: 'left',
-          height: '14px',
-          width: '14px'
-        });
-        promptNode.appendChild(iconNode);
-      }
-    };
-
-    const toggleCellMetadata = (
-      editable: boolean,
-      deletable: boolean,
-      tracker: INotebookTracker
-    ) => {
-      const current = tracker.currentWidget;
-      if (!current) {
-        console.warn('No active notebook.');
-        return;
-      }
-
-      const notebook = current.content;
-      const cells = notebook.model?.cells;
-      if (!cells) {
-        return;
-      }
-
-      let editedCellCount = 0;
-      let nonEditedCellCount = 0;
-      for (let i = 0; i < cells.length; i++) {
-        const cellModel = cells.get(i);
-        const cellWidget = notebook.widgets[i];
-
-        const isEditable = asBool(cellModel.getMetadata('editable'));
-        const isDeletable = asBool(cellModel.getMetadata('deletable'));
-
-        if (isEditable !== editable || isDeletable !== deletable) {
-          cellModel.setMetadata('editable', editable);
-          cellModel.setMetadata('deletable', deletable);
-
-          applyCellLockIcon(cellModel, cellWidget);
-          editedCellCount++;
-        } else {
-          nonEditedCellCount++;
-        }
-      }
-
-      const action = editable ? 'unlocked' : 'locked';
-      let statusMessage = '';
-      if (editedCellCount === 0) {
-        statusMessage = `All cells were already ${action}.`;
-      } else {
-        statusMessage = `${editedCellCount} cell${editedCellCount > 1 ? 's' : ''} ${
-          editedCellCount > 1 ? 'were' : 'was'
-        } successfully ${action}.`;
-        if (nonEditedCellCount > 0) {
-          statusMessage += ` (${nonEditedCellCount} already ${action}).`;
-        }
-      }
-
-      // Show message in status bar transiently
-      if (statusWidget) {
-        statusWidget.setTemporaryStatus(statusMessage);
-      } else {
-        console.log('[CellLockStatus]', statusMessage);
-      }
-    };
-
     // Define the lock command
     const lockCommand = 'jupyterlab-cell-lock:lock-cells';
     app.commands.addCommand(lockCommand, {
       label: 'Make All Current Cells Read-Only & Undeletable',
       execute: () => {
-        toggleCellMetadata(false, false, tracker);
+        toggleCellMetadata(false, false, tracker, statusWidget);
       }
     });
 
@@ -197,34 +46,9 @@ const plugin: JupyterFrontEndPlugin<void> = {
     app.commands.addCommand(unlockCommand, {
       label: 'Make All Current Cells Editable & Deletable',
       execute: () => {
-        toggleCellMetadata(true, true, tracker);
+        toggleCellMetadata(true, true, tracker, statusWidget);
       }
     });
-
-    const refreshLockIcons = (notebookPanel: any) => {
-      if (!notebookPanel) {
-        return;
-      }
-      const { content: notebook } = notebookPanel;
-
-      // Ensure the notebook model is ready before trying to access cells
-      if (notebook.model && notebook.widgets) {
-        console.log(
-          'Refreshing lock icons for',
-          notebook.widgets.length,
-          'cells'
-        );
-
-        requestAnimationFrame(() => {
-          notebook.widgets.forEach((cellWidget: any, i: number) => {
-            const cellModel = notebook.model.cells.get(i);
-            if (cellModel) {
-              applyCellLockIcon(cellModel, cellWidget);
-            }
-          });
-        });
-      }
-    };
 
     tracker.widgetAdded.connect((_, notebookPanel) => {
       const { content: notebook, context } = notebookPanel;
